Fix cherry-picking Active Support core extension
As suggested in the guide, we first need to require 'active_support',
(supposed to have near-zero footprint) - without it other gems may
incorrectly assume it can access methods on the ActiveSupport module, for
instance .on_load and this will result in a NoMethodError.

The awesome_print gem is one such example.