Serialize to JSON when content type is set to JSON
Until now, the request body is only serialized if it's a Hash.
But it could be an Array, or any other serializable type,
and those bodies are serialized. See #129, probably #141 as well.

I am not sure if this wouldn't break use cases where
the body is already expected to be serialized.